Render compatibility tables properly, put page move JS into own file Removes HTML5 required attribute from content block
