+++ apps/wiki/templates/wiki/includes/tag_suggestions.html
@@ -1,3 +1,3 @@
 <script type="text/javascript">
-  mdn.wiki.tagSuggestions = {{ WIKI_DOCUMENT_TAG_SUGGESTIONS|safe }};
+  mdn.wiki.tagSuggestions = {{ WIKI_DOCUMENT_TAG_SUGGESTIONS|safe }} || [];
 </script>
