+++ apps/wiki/templates/wiki/move_document.html
@@ -87,3 +87,7 @@
    </div>
 </section>
 {% endblock %}
+
+{% block js %}
+  <script src="{{ MEDIA_URL }}js/wiki-move.js?build={{ BUILD_ID_JS }}"></script>
+{% endblock %}
