+++ settings.py
@@ -668,9 +668,10 @@ MINIFY_BUNDLES = {
             'js/libs/jquery-ui-1.10.3.custom/js/jquery-ui-1.10.3.custom.min.js',
             'js/moz-jquery-plugins.js',
             'js/main.js',
+            'js/wiki.js',
         ),
         'wiki-edit': (
-            'js/wiki.js',
+            'js/wiki-edit.js',
             'js/libs/tag-it.js',
             'js/wiki-tags-edit.js',
         ),
